Updated certain format placeholders which did not work well on 64-bit systems.

+++ framework/Source/CPTGradient.m
@@ -700,7 +700,7 @@ static void resolveHSV(CGFloat *color1, CGFloat *color2);
     CPTGradientElement removedElement = [newGradient removeElementAtIndex:index];
 
     if ( isnan(removedElement.position) ) {
-        [NSException raise:NSRangeException format:@"-[%@ removeColorStopAtIndex:]: index (%i) beyond bounds", [self class], index];
+        [NSException raise:NSRangeException format:@"-[%@ removeColorStopAtIndex:]: index (%li) beyond bounds", [self class], index];
     }
 
     return [newGradient autorelease];
@@ -727,7 +727,7 @@ static void resolveHSV(CGFloat *color1, CGFloat *color2);
 #endif
     }
 
-    [NSException raise:NSRangeException format:@"-[%@ colorStopAtIndex:]: index (%i) beyond bounds", [self class], index];
+    [NSException raise:NSRangeException format:@"-[%@ colorStopAtIndex:]: index (%li) beyond bounds", [self class], (long)index];
 
     return NULL;
 }
